Adds StyleFactory.
